requirements.txt
numpy==1.24.4
pandas==2.0.3
typeguard==4.3.0
Jinja2==3.1.4
PyYAML==6.0.1
openpyxl==3.1.3

